+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.19.1...main`][0.19.1...main].
+For a full diff see [`0.20.0...main`][0.20.0...main].
+
+## [`0.20.0`][0.20.0] - 2026-04-10
+
+For a full diff see [`0.19.1...0.20.0`][0.19.1...0.20.0].
 
 ### Added
 
@@ -382,6 +386,7 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.18.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.18.0
 [0.19.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.19.0
 [0.19.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.19.1
+[0.20.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.20.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
 [0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
@@ -404,4 +409,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.17.0...0.18.0]: https://github.com/weph/obsidian-bookshelf/compare/0.17.0...0.18.0
 [0.18.0...0.19.0]: https://github.com/weph/obsidian-bookshelf/compare/0.18.0...0.19.0
 [0.19.0...0.19.1]: https://github.com/weph/obsidian-bookshelf/compare/0.19.0...0.19.1
-[0.19.1...main]: https://github.com/weph/obsidian-bookshelf/compare/0.19.1...main
+[0.19.1...0.20.0]: https://github.com/weph/obsidian-bookshelf/compare/0.19.1...0.20.0
+[0.20.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.20.0...main